Update CumSumSuite.R
Quick fix for cumsumbelow, which checks the `vals` object to see if it's numeric but the vals object is a matrix. So now it checks the data in the first row. This only works in the matrix is all numeric, but since it's coming from a raster layer that should be the case (i.e., all data should be the same type).